refact: rename types.go file name
